fixed all instances of JWT_SECRET
In a previous project I had put "JWT_Secret" instead of "JWT_SECRET_KEY" and as a result the login didn't work. Took me a long time to figure out what was wrong. I noticed it in this project so I went ahead and changed them so we don't run into the same error.